\name{dataexample}
\alias{dataexample}
\alias{datex}
\title{Stock market indexes (daily closing prices).
}
\description{
Seven European stock market indexes (daily closing prices): MIB30 (Italy),
IBEX35 (Spain), DAX30 (Germany), CAC40 (France), AEX25 (Netherlands), 
ATX20 (Austria) and NBEL20 (Belgium) spanning from January 2, 2004 to 
June 29, 2012. 
}
\usage{data(dataexample)
}
\format{A list containing 2216 elements and 7 variables
}
\source{http://finance.yahoo.com
}
\keyword{datasets}

